added another model for fuel and finished collision system

+++ source/common/components/car.hpp
@@ -7,8 +7,8 @@
 namespace our {
     class CarComponent : public Component {
     public:
-        int width = 0;
-        int depth = 0;
+        float width = 0;
+        float depth = 0;
         int health = 100;
 
         // The ID of this component type is "Car"
